requirements.txt
alabaster==0.7.12
asgiref==3.2.7
astroid==2.3.3
attrs==19.3.0
Babel==2.8.0
backcall==0.1.0
beautifulsoup4==4.9.1
bleach==3.1.0
certifi==2020.4.5.2
cffi==1.14.0
chardet==3.0.4
colorama==0.4.3
cryptography==2.9.2
cycler==0.10.0
decorator==4.4.1
defusedxml==0.6.0
Django==2.1.12
django-allauth==0.42.0
django-bootstrap-pagination==1.7.1
django-bootstrap4==1.1.1
django-cors-headers==3.3.0
django-extensions==2.2.9
django-rest-auth==0.9.5
djangorestframework==3.11.0
docutils==0.16
entrypoints==0.3
idna==2.9
imagesize==1.2.0
importlib-metadata==1.6.1
ipykernel==5.1.4
ipython==7.11.1
ipython-genutils==0.2.0
ipywidgets==7.5.1
isort==4.3.21
jedi==0.16.0
Jinja2==2.11.0
json5==0.8.5
jsonschema==3.2.0
jupyter==1.0.0
jupyter-client==5.3.4
jupyter-console==6.1.0
jupyter-core==4.6.1
jupyterlab==1.2.6
jupyterlab-server==1.0.6
kiwisolver==1.1.0
lab==5.5
lazy-object-proxy==1.4.3
MarkupSafe==1.1.1
matplotlib==3.2.1
mccabe==0.6.1
mistune==0.8.4
nbconvert==5.6.1
nbformat==5.0.4
notebook==6.0.3
numpy==1.18.2
oauthlib==3.1.0
packaging==20.4
pandocfilters==1.4.2
parso==0.6.0
pickleshare==0.7.5
prometheus-client==0.7.1
prompt-toolkit==3.0.3
pycparser==2.20
Pygments==2.5.2
PyJWT==1.7.1
pylint==2.4.4
pyparsing==2.4.6
pyrsistent==0.15.7
python-dateutil==2.8.1
python-decouple==3.3
python3-openid==3.1.0
pytz==2020.1
pywinpty==0.5.7
pyzmq==18.1.1
qtconsole==4.7.2
QtPy==1.9.0
requests==2.23.0
requests-oauthlib==1.3.0
Send2Trash==1.5.0
simplejson==3.17.0
six==1.15.0
snowballstemmer==2.0.0
social-auth-app-django==3.4.0
social-auth-core==3.3.3
soupsieve==2.0.1
Sphinx==3.1.1
sphinx-rtd-theme==0.4.3
sphinxcontrib-applehelp==1.0.2
sphinxcontrib-devhelp==1.0.2
sphinxcontrib-htmlhelp==1.0.3
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==1.0.3
sphinxcontrib-serializinghtml==1.1.4
sqlparse==0.3.1
terminado==0.8.3
testpath==0.4.4
tornado==6.0.3
traitlets==4.3.3
txt2tags==3.7
typed-ast==1.4.1
urllib3==1.25.9
wcwidth==0.1.8
webencodings==0.5.1
widgetsnbextension==3.5.1
wrapt==1.11.2
zipp==3.1.0
